Buoi6/requirements.txt
lxml==5.2.2
mysql-connector-python==9.0.0
